+++ jaot_base/tests/__init__.py
@@ -6,3 +6,4 @@ from . import test_lifecycle
 from . import test_security
 from . import test_error_paths
 from . import test_adversarial
+from . import test_cases